Move contextlib.chdir to shutil.chdir, replacing chdir with a compatibility function in contextlib, which has a deprecation warning. Main tests for chdir moved to test_shutil, adding a new test to test_contextlib to test the compatibility function
